requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0
PyJWT==2.8.0
APScheduler==3.10.4
